compositing: Move WebRender initialization to `IOCompositor` creation
In order to support multiple WebRender instances in the renderer, we
need the ability for the renderer itself to manage their creation and
destruction. This change moves the initialization of the (currently)
single WebRender instances to the renderer initialization.
